WIP Viz <=> Tests refactor: move (de)serialization code into viz.rs
